Move gce-machines to src/analysis
