+++ src/analysis/RuntimeManager.js
@@ -6,11 +6,11 @@ import { Fragment, useEffect, useState } from 'react';
 import { div, h, p } from 'react-hyperscript-helpers';
 import { appLauncherTabName, GalaxyLaunchButton, GalaxyWarning } from 'src/analysis/runtime-common-components';
 import { getCurrentApp } from 'src/analysis/utils/app-utils';
+import { dataSyncingDocUrl } from 'src/analysis/utils/gce-machines';
 import { cloudProviders, getCurrentRuntime } from 'src/analysis/utils/runtime-utils';
 import { appTools, runtimeToolLabels } from 'src/analysis/utils/tool-utils';
 import { ButtonPrimary, Clickable, Link, spinnerOverlay } from 'src/components/common';
 import Modal from 'src/components/Modal';
-import { dataSyncingDocUrl } from 'src/data/gce-machines';
 import { Ajax } from 'src/libs/ajax';
 import { getDynamic, getSessionStorage, setDynamic } from 'src/libs/browser-storage';
 import colors from 'src/libs/colors';
