+++ src/analysis/modal-utils.ts
@@ -1,7 +1,7 @@
 import _ from 'lodash/fp';
+import { cloudServices } from 'src/analysis/utils/gce-machines';
 import { defaultDataprocMachineType, getDefaultMachineType } from 'src/analysis/utils/runtime-utils';
 import { getToolLabelFromCloudEnv } from 'src/analysis/utils/tool-utils';
-import { cloudServices } from 'src/data/gce-machines';
 import { GooglePdType, PersistentDisk, SharedPdType } from 'src/libs/ajax/leonardo/models/disk-models';
 
 export interface IComputeConfig {
