+++ src/analysis/modals/DeleteDiskChoices.test.ts
@@ -4,9 +4,9 @@ import { h } from 'react-hyperscript-helpers';
 import { azureRuntime, getAzureDisk, getDisk, getRuntimeConfig } from 'src/analysis/_testData/testData';
 import { DeleteDiskChoices } from 'src/analysis/modals/DeleteDiskChoices';
 import { DeleteEnvironment } from 'src/analysis/modals/DeleteEnvironment';
+import { cloudServices } from 'src/analysis/utils/gce-machines';
 import { runtimeToolLabels } from 'src/analysis/utils/tool-utils';
 import { ButtonPrimary } from 'src/components/common';
-import { cloudServices } from 'src/data/gce-machines';
 import { cloudServiceTypes } from 'src/libs/ajax/leonardo/models/runtime-config-models';
 import { formatUSD } from 'src/libs/utils';
 import { cloudProviderTypes } from 'src/libs/workspace-utils';
