+++ src/analysis/utils/gce-machines.js
